pin version of sphinxcontrib-bibtex<2.0.0
The root cause of previous build failures seems to be related to an
incompatible upgrade of the `sphinxcontrib-bibtex` package as also noted
in executablebooks/jupyter-book#1137. This pin can most likely be
removed once a new version of `jupyter-book` is available, either
because `jupyter-book` includes the pin in its dependencies or because it
adapts to the new requirements of `sphinxcontrib-bibtex`.

I've also removed the tag for the gitlab runner as it is very likely
that it was working on the one runner and not on the other because it
cached an older version of `sphinxcontrib-bibtex` and as soon as the
cache would be emptied the issue might arise again.