add interactive flight track maps
This notebook creates an interactive view on HALO's flight tracks. I
wouldn't consider this as an entry level notebook, but I think it is
still not too complicated. On the plus side, the final map looks quite
cool and I think the whole code doesn't require any additional binary
packages.