increase cell runtime limit
The cell runtime limit shouldn't grow to such large values. However, for
some reason, loading datasets concurrently doesn't work (see #58). In
the long run, we should find a way to solve the referenced issue and
revert this commit, but for now, I can't find a better alternative and
not fixing this will likely result in many failed builds in the near
future.